market-aggregation-service/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
cryptography==41.0.7
pyjwt==2.8.0
python-dateutil==2.8.2
pytz==2023.3

